new def of TP, remove BIC, reformatting
I thought we would want to compare CCA, so I changed all the functions
to include CCA output.  Then I realized that the whole point of SCCA was
to look at the sparsity (which we can't measure with CCA).  So I
completely removed all the CCA stuff.  But a lot of the variable names
changed (mostly to now include the prefix "sp."), so make sure to use
these updated functions.